+++ server.js
@@ -4,7 +4,7 @@ const express = require('express');
 const sequelize = require('./config/connection');
 
 //Import all routes
-const routes = require('./routes');
+const routes = require('./controllers');
 
 
 const app = express();
@@ -16,12 +16,11 @@ app.use(express.urlencoded({ extended: true }));
 // turn on routes
 app.use(routes);
 
-
 // turn on connection to db and server
 const init = async () => {
   try {
     await sequelize.sync({force:true});
-    console.log('Sucessful connection to the database');
+    console.log('Successful connection to the database');
     app.listen(PORT, () => console.log('Express web server now listening'));
   } catch (err) {
     console.log(err);
